requirements.txt
# Core ML
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0

# Logging & Config
wandb>=0.16.0
pyyaml>=6.0
omegaconf>=2.3.0

# Utilities
tqdm>=4.65.0
numpy>=1.24.0

# Testing
pytest>=7.4.0


hf_transfer==0.1.9